validationshipping

+++ server/services/product.ts
@@ -2,41 +2,60 @@ import {Strapi} from '@strapi/strapi';
 
 export default ({ strapi }: { strapi: Strapi }) => ({
 
-  async find(query) {
-    return await strapi.entityService.findMany("plugin::omcommerce.product", {
-      fields: [
-        'id',
-        'description',
-        'amount_currency_code',
-        'amount_value',
-        'amount_value_rsd',
-        'Quantity',
-        'weight',
-        'measurement_unit',
-        'title',
-        'slug',
-      ],
-      populate: {media: true, categories: true},
-      filters:
-      query.filters
-    });
-
+  async find(query: any) {
+    if (strapi.entityService) {
+      return await strapi.entityService.findMany("plugin::omcommerce.product", {
+        fields: [
+          'id',
+          'description',
+          'amount_currency_code',
+          'amount_value',
+          'amount_value_converted',
+          'Quantity',
+          'weight',
+          'measurement_unit',
+          'title',
+          'slug',
+        ],
+        populate: {media: true, categories: true},
+        filters:
+        query.filters
+      });
+    } else {
+      throw new Error('strapi.entityService is not defined');
+    }
   },
 
-  async findOne(id) {
-    return await strapi.entityService.findOne("plugin::omcommerce.product", id);
+  async findOne(id: any) {
+    if (strapi.entityService) {
+      return await strapi.entityService.findOne("plugin::omcommerce.product", id);
+    } else {
+      throw new Error('strapi.entityService is not defined');
+    }
   },
 
-  async create(data) {
-    return await strapi.entityService.create("plugin::omcommerce.product", data);
+  async create(data: any) {
+    if (strapi.entityService) {
+      return await strapi.entityService.create("plugin::omcommerce.product", data);
+    } else {
+      throw new Error('strapi.entityService is not defined');
+    }
   },
 
-  async update(id, data) {
-    return await strapi.entityService.update("plugin::omcommerce.product", id, data);
+  async update(id: any, data: any) {
+    if (strapi.entityService) {
+      return await strapi.entityService.update("plugin::omcommerce.product", id, data);
+    } else {
+      throw new Error('strapi.entityService is not defined');
+    }
   },
 
-  async delete(id) {
-    return await strapi.entityService.delete("plugin::omcommerce.product", id);
+  async delete(id: any) {
+    if (strapi.entityService) {
+      return await strapi.entityService.delete("plugin::omcommerce.product", id);
+    } else {
+      throw new Error('strapi.entityService is not defined');
+    }
   },
 
 });
